backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
pgvector==0.2.4
sqlalchemy==2.0.23
sentence-transformers==2.2.2
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
bcrypt==4.1.2
pyjwt==2.8.0
uuid==1.30
bleach==6.1.0
